+++ src/main/java/com/back/b2st/domain/reservation/controller/SeatSelectionController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.back.b2st.domain.reservation.service.SeatSelectionService;
+import com.back.b2st.domain.reservation.service.SeatHoldService;
 import com.back.b2st.global.common.BaseResponse;
 
 import lombok.RequiredArgsConstructor;
@@ -15,14 +15,14 @@ import lombok.RequiredArgsConstructor;
 @RequestMapping("/schedules/{scheduleId}/seats")
 public class SeatSelectionController {
 
-	private final SeatSelectionService seatSelectionService;
+	private final SeatHoldService seatHoldService;
 
 	@PostMapping("/{seatId}/hold")
 	public BaseResponse<Void> holdSeat(
 		@PathVariable Long scheduleId,
 		@PathVariable Long seatId
 	) {
-		seatSelectionService.holdSeat(scheduleId, seatId);
+		seatHoldService.holdSeat(scheduleId, seatId);
 		return BaseResponse.success();
 	}
